+++ e2e_playwright/canvas_background_color.py
@@ -28,7 +28,7 @@ st.subheader("initial_drawing round-trip")
 result_round_trip = st_canvas(
     height=200,
     width=300,
-    drawing_mode="transform",
+    drawing_mode="edit",
     initial_drawing=result.json_data,
     key="canvas2",
 )