+++ streamlit_drawable_canvas/frontend/src/tools/edit.ts
@@ -1,6 +1,6 @@
 import { FabricTool, ConfigureCanvasProps } from "./fabrictool";
 
-export class TransformTool extends FabricTool {
+export class EditTool extends FabricTool {
   configureCanvas(_args: ConfigureCanvasProps): () => void {
     const canvas = this.canvas;
     canvas.isDrawingMode = false;